refactor(headers): split marufs.h / marufs_layout.h into per-domain headers
marufs.h was a 1266-line catch-all (sb_info + DAX/RAT/shard helpers +
function decls for 10 modules). marufs_layout.h was a 625-line mix of
on-disk structs and CXL/CAS primitives. Both now act as umbrella
headers over focused per-domain files.

Phase 1 — extract reusable primitives + per-module decl headers:
  marufs_endian.h  READ_LE/WRITE_LE/READ_CXL_LE, MARUFS_CXL_WMB/RMB,
                   le16/32/64_cas, cas_inc/dec
  marufs_hash.h    shard_idx, bucket_idx, hash_name, make_ino,
                   ino_to_region, align_up
  gc.h             orphan tracker types + gc.c entry points
  inode.h          marufs_inode_info struct + inode.c entry points
                   + inode_ops externs
  acl.h, cache.h, dir.h, file.h, index.h, nrht.h, region.h, super.h
                   per-module function declarations

Phase 2 — split on-disk structs by subsystem:
  marufs_superblock_layout.h   marufs_superblock + GSB_SIZE
  marufs_index_layout.h        shard_header, index_entry, region
                               defaults, BUCKET_END, state enum
  marufs_rat_layout.h          rat, rat_entry, deleg_entry, RAT/deleg/
                               region_type state enums, capacity
  marufs_nrht_layout.h         nrht_header, nrht_shard_header,
                               nrht_entry, NRHT defaults

Umbrella files now hold only:
  marufs.h          sb_info, DAX/RAT/shard inline accessors, sysfs decls
  marufs_layout.h   magic enum, ME area sizes + me_area_size helper,
                    layout offsets, compile-time size validators

Existing .c files keep including marufs.h alone — umbrella pulls in
all per-module headers, so include patterns are unchanged. Phase 3
(sb_info field grouping into sub-structs) deferred — too invasive,
low ROI.

Line counts:
  marufs.h         1266 -> 435  (66% reduction)
  marufs_layout.h   625 -> 130  (79% reduction)
  16 new headers    ~970 lines  (focused, dependency-minimal)

Build clean. Compile-time size validators still pass.
