+++ marufs_kernel/src/dir.h
@@ -1,0 +1,13 @@
+/* SPDX-License-Identifier: GPL-2.0-only */
+/*
+ * dir.h - VFS directory file_operations export.
+ */
+
+#ifndef _MARUFS_DIR_H
+#define _MARUFS_DIR_H
+
+#include <linux/fs.h>
+
+extern const struct file_operations marufs_dir_ops;
+
+#endif /* _MARUFS_DIR_H */
